% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adssn.R
\docType{package}
\name{adssn}
\alias{adssn}
\alias{adssn-package}
\title{adssn: Adaptive Design for Spatial Stream Networks}
\description{
The adssn package builds upon the SSN package to provide a complete workflow for adaptive sampling design for models of spatial stream networks.
}
\section{adssn functions}{

the adssn package contains the following functions: (list of function names with brief description and link to associated manual pages)
}

