% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandaid_fix.R
\name{bandaid_fix}
\alias{bandaid_fix}
\title{A wrapper function for SimulateOnSSN that avoids issues with the use of Random Effects.}
\usage{
bandaid_fix(...)
}
\description{
A wrapper function for SimulateOnSSN that avoids issues with the use of Random Effects.
}
