% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateShreveStreamOrderAndAFVs.R
\name{calculateShreveStreamOrderAndAFVs}
\alias{calculateShreveStreamOrderAndAFVs}
\title{Calculate Shreve stream order and afv}
\usage{
\code{calculateShreveStreamOrderAndAFVs(ssn)}
}
\arguments{
\item{ssn}{an object of class SpatialStreamNetwork}

\item{BID.tables}{optionally provide pre-computed binary ID tables. This can be beneficial if the stream network is large or if the binary ID tables have already been computed for another purpose.}
}
\value{
An object of class SpatialStreamNetwork.
}
\description{
Calculates Shreve stream order and additive function values on an arbitrary SpatialStreamNetwork object.
}
\details{
This function accepts a SpatialStreamNetwork object and calculates Shreve stream orders and additive function values for it, which are then joined back to the `ssn@data` slot. This funciton should be used when a SpatialStreamNetwork has no edge attributes which could be used to calculate the AFV.
}
