% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseAndReorderEdgelist.R
\name{collapseAndReorderEdgelist}
\alias{collapseAndReorderEdgelist}
\title{A function to crush down an edgelist in to a vector of paired vertices}
\usage{
\code{collapseAndReorderEdgelist(edgelist)}
}
\arguments{
\item{edgelist}{An n x 2 matrix, as produced by the function \code{get.edgelist()}}
}
\value{
A numeric vector.
}
\description{
An edgelist is an n x 2 matrix which details connections between edges in a network; from the first column to the second column. This function reduces the matrix to a vector which is more convenient to deal with inside the function \code{generateSites}.
}
\details{
This is a function called inside \code{generateSites} to correctly place simulated sites on a SpatialStreamNetwork. It is not intended for use outside of this context, though it may prove useful in other applications.
}
