% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructTotalMatrix.R
\name{constructTotalMatrix}
\alias{constructTotalMatrix}
\title{Construct distance and net.zero matrices for a SpatialStreamNetwork with multiple networks}
\usage{
\code{constructTotalMatrix(list.matrices)}
}
\arguments{
\item{list.matrices}{A list whose elements are the dist.junc matrices for each network. The networks are assumed to be in ascending numerical order.}

\item{pxo}{A logical indicating whether the function is constructing the dist.junc matrix between the observed and predicted sites.}
}
\value{
A list of two matrices: one being the total distance matrix and the other being a binary matrix referred to as net.zero.
}
\description{
This function combines the dist.junc matrices from each network into a single matrix.
}
\details{
This function is used internally in \code{\link{findOptimalDesign}} and \code{\link{doAdaptiveDesign}}.
}
