% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPOptimality.R
\name{CPOptimality}
\alias{CPOptimality}
\title{A utility function for covariance parameter estimation}
\usage{
\code{CPOptimality(ssn, glmssn, design.points, prior.parameters, n.draws, extra.arguments)}
}
\arguments{
\item{ssn}{An object of class SpatialStreamNetwork}

\item{glmssn}{An model object of class glmssn.}

\item{design.points}{A vector of pids corresponding to a set of observed sites in the obspoints slot of the SpatialStreamNetwork object.}

\item{prior.parameters}{A list of random functions that are parameterised in terms of n.draws.}

\item{n.draws}{A numeric scalar for the number of Monte Carlo draws to use when approximating the utility.}

\item{extra.arguments}{A list of extra parameters that control the behaviour of the utility function. The distance matrices required to compute covariance matrices are also stored in this list. Note that these are generated inside \code{\link{findOptimalDesign}} and \code{\link{doAdaptiveDesign}}.}
}
\value{
A numeric scalar.
}
\description{
\code{CPOptimality} is a utility function that can be used with \code{\link{findOptimalDesign}}. It is a utility function that minimises the determinant of the inverse Fisher information matrix over a set of designs.
}
\details{
\code{CPOptimality} assumes there are no sites which have already been incorporated into a design. It computes the expected Fisher information matrix (Som et al., 2014) for each set of simulated covariance parameters.
}
