% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBinaryID.R
\name{createBinaryID}
\alias{createBinaryID}
\title{Create binaryID tables for a SpatialStreamNetwork}
\usage{
createBinaryID(ssn, o.write)
}
\arguments{
\item{ssn}{an object of class SpatialStreamNetwork}

\item{o.write}{a logical value}
}
\value{
None.
}
\description{
\code{createBinaryID()} generates binaryID tables in a binaryID database inside a .ssn folder.
}
\section{Warning}{

This function is not intended for direct use.
}

