% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOptimality.R
\name{DOptimality}
\alias{DOptimality}
\title{A utility function for fixed effects parameter estimation with known covariance parameters}
\usage{
\code{DOptimality(ssn, glmssn, design.points, prior.parameters, n.draws, extra.arguments)}
}
\arguments{
\item{ssn}{An object of class SpatialStreamNetwork}

\item{glmssn}{An model object of class glmssn.}

\item{design.points}{A vector of pids corresponding to a set of observed sites in the obspoints slot of the SpatialStreamNetwork object.}

\item{prior.parameters}{A list of random functions that are parameterised in terms of n.draws.}

\item{n.draws}{A numeric scalar for the number of Monte Carlo draws to use when approximating the utility.}

\item{extra.arguments}{A list of extra parameters that control the behaviour of the utility function. The distance matrices required to compute covariance matrices are also stored in this list. Note that these are generated inside \code{\link{findOptimalDesign}} and \code{\link{doAdaptiveDesign}}.}
}
\value{
A numeric scalar.
}
\description{
\code{DOoptimality} is a utility function that can be used with \code{\link{findOptimalDesign}}. It is a utility function that minimises the determinant of the variance-covariance matrix of the fixed effects over a set of designs.
}
\details{
This utility function is similar to \code{\link{EDOptimality}}, except that it assumes the covariance parameters are known. The definition of this utility function follows that in Som et al. (2014), with a minor modification to allow the incorporation of prior information in the form of Monte Carlo simulations.
}
