% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawStreamNetworkSamples.R
\name{drawStreamNetworkSamples}
\alias{drawStreamNetworkSamples}
\title{Construct probability-based and heuristic designs from Som et al. (2014).}
\usage{
drawStreamNetworkSamples(ssn, new.ssn.path, overwrite.path = FALSE,
  sample.method, sample.size, use.locID = FALSE, ...)
}
\arguments{
\item{ssn}{An object of class SpatialStreamNetwork.}

\item{new.ssn.path}{A path for the new .ssn directory where the results should be written out.}

\item{overwrite.path}{A logical indicating whether the ssn directory referred to in new.ssn.path should be overwritten, if it exists already. Defaults to FALSE.}

\item{sample.method}{A character vector providing the label for the specific sampling design that is desired, with references to Som et al. (2014) including simple random sample "SRS", "GRTS", "GRTSmouth", "GRTSclus","Extreme.Clust.and.Singles", "Trib.Sets.Ext.Singles.sample", "Trib.Sets.Ext.Singles.Mouth.sample"}

\item{sample.size}{A numeric scalar specifying the desired sample size.}

\item{use.locID}{A logical indicating whether sampling sites should be selected by locID instead of pid. Defaults to FALSE.}

\item{...}{Other arguments to Stream.Network.Samples; chiefly, cluster.number, which is required for some sample.method arguments.}
}
\value{
An object of class SpatialStreamNetwork. Note, any prediction points will have to be imported separately.
}
\description{
This function is a wrapper for \code{\link{Stream.Network.Samples}}. This function constructs any of the designs outlined in Som et al. (2014), such as GRTS and stream network designs with clusters around confluences.
}
\references{
Som, N.A., Monestiez, P., Ver Hoef, J.M., Zimmerman, D.L., & Peterson, E.E. (2014). Spatial sampling on streams: principles for inference on aquatic networks. \emph{Environmetrics}, \emph{25}(5), 306-323. doi: 10.1002/env.2284.
}
