% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDOptimality.R
\name{EDOptimality}
\alias{EDOptimality}
\title{A utility function for optimal designs for the empirical estimation of fixed effects parameters.}
\usage{
\code{EDOptimality(ssn, glmssn, design.points, prior.parameters, n.draws, extra.arguments)}
}
\arguments{
\item{ssn}{An object of class SpatialStreamNetwork}

\item{glmssn}{An model object of class glmssn.}

\item{design.points}{A vector of pids corresponding to a set of observed sites in the obspoints slot of the SpatialStreamNetwork object.}

\item{prior.parameters}{A list of random functions that are parameterised in terms of n.draws.}

\item{n.draws}{A numeric scalar for the number of Monte Carlo draws to use when approximating the utility.}

\item{extra.arguments}{A list of extra parameters that control the behaviour of the utility function. The distance matrices required to compute covariance matrices are also stored in this list. Note that these are generated inside \code{\link{findOptimalDesign}} and \code{\link{doAdaptiveDesign}}.}
}
\value{
A numeric scalar.
}
\description{
\code{EDOoptimality} is a utility function that can be used with \code{\link{findOptimalDesign}}. It is a utility function that minimises the determinant of the variance-covariance matrix of the fixed effects over a set of designs.
}
\details{
This utility function is similar to \code{\link{DOptimality}}, except that it does not assume the covariance parameters are known. These are instead estimated from the data associated with a set of design points, while still incorporating prior information about the covariance parameters.
}
