% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolveRandOverTime.R
\name{evolveRandOverTime}
\alias{evolveRandOverTime}
\title{Constructing random designs over time}
\usage{
evolveRandOverTime(ssn, num.sites, rep.variable)
}
\arguments{
\item{ssn}{A \code{SpatialStreamNetwork} object}

\item{num.sites}{A numeric vector. This vector can contain a single element or as many elements as there are temporal steps in the \code{ssn} sites}

\item{rep.variable}{A string being the variable on which the temporal replicates are defined}
}
\value{
A list which indicates the design for each timestep. The designs are given in lists with two named elements: \code{by.locID}, where the sites are given by their locID values; \code{by.pid}, where the sites are given by their pid values.
}
\description{
This function takes a \code{SpatialStreamNetwork} object with temporally replicated sites and returns a random design that evolves over time. Once sites are included in the design, they remain for all subsequent timesteps.
}
