% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractExpectedFisherInformation.R
\name{extractExpectedFisherInformation}
\alias{extractExpectedFisherInformation}
\title{Extract the expected Fisher information matrix for the covariance parameters in a glmssn object}
\usage{
extractExpectedFisherInformation(glmssn,
  covparms = glmssn$estimates$theta, diff = 1e-08)
}
\arguments{
\item{glmssn}{An object of class \code{glmssn}.}

\item{covparms}{A vector of covariance parameter values. These are extracted from the glmssn object by default.}

\item{diff}{The step size used for finite differencing.}
}
\value{
The expected Fisher information matrix.
}
\description{
This function takes a \code{glmssn} object and returns the expected Fisher information matrix on its covariance parameters.
}
