% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.seg.below.R
\name{find.seg.below}
\alias{find.seg.below}
\title{Find the rid for the reach immediatley downstream of another reach.}
\usage{
find.seg.below(x, IDtable)
}
\arguments{
\item{x}{a numeric scaler which represents the binaryID address for a rich object of class data.frame which is a single reach's subset of the point.data slot fomr a SpatialStreamNetwork}

\item{IDtable}{an object of class data.frame which represents the bindaryID object for a SpatialStreamNetwork}
}
\value{
a numeric scalar which is the rid for the reach immediatley downstream of the reach whose address is provided as x
}
\description{
For a reach in a SpatialStreamNetwork, this function finds the rid for the reach immediately downstream.
}
