% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.segment.lengths.R
\name{find.segment.lengths}
\alias{find.segment.lengths}
\title{Find the length of every reach in a SpatialStreamNetwork.}
\usage{
find.segment.lengths(raw.ssn, bin.table)
}
\arguments{
\item{raw.ssn}{an object of class SpatialStreamNetwork}

\item{bin.table}{an object of class data.frame which represents the bindaryID object for a SpatialStreamNetwork}
}
\value{
a data.frame
}
\description{
For a SpatialStreamNetwork, find the length of all reaches.
}
