% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Find.Tribs.R
\name{Find.Tribs}
\alias{Find.Tribs}
\title{For each tributary junction in a SpatialStreamNetwork, find the rids of all reaches that meet to form the tributary junction.}
\usage{
Find.Tribs(ssn.obj, bin.table)
}
\arguments{
\item{ssn.obj}{an object of class SpatialStreamNetwork}

\item{bin.table}{an object of class data.frame which represents the bindaryID object for a SpatialStreamNetwork}
}
\value{
a data.frame
}
\description{
For each tributary junction in a SpatialStreamNetwork, find the rids of all reaches that meet to form the tributary junction.
}
