% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findOptimalDesign.R,
%   R/findOptimalDesignParallel.R
\name{findOptimalDesign}
\alias{findOptimalDesign}
\alias{findOptimalDesign}
\title{Implement optimal designs on a SpatialStreamNetwork}
\usage{
\code{findOptimalDesign(ssn, glmssn, afv.column, n.points, utility.function, prior.parameters, n.draws = 500, extra.arguments = NULL)}

\code{findOptimalDesign(ssn, glmssn, afv.column, n.points, utility.function, prior.parameters, n.draws = 500, extra.arguments = NULL)}
}
\arguments{
\item{ssn}{an object of class SpatialStreamNetwork}

\item{glmssn}{an object of class glmssn}

\item{afv.column}{the name of the column in the SpatialStreamNetwork object that contains the additive function values}

\item{n.points}{the number of points to be included in the final design. This can be a single number, in which case all networks will have the same number of points. This can also be a vector with the same length as the number of networks in ssn.}

\item{utility.function}{a function with the signature utility.function. Users may define their own. This package provides several built-in utility functions, including \code{\link{DOptimality}}, \code{\link{KOptimality}}, and \code{\link{CPOptimality}}.}

\item{prior.parameters}{a function to act as a prior for covariance parameter values}

\item{n.draws}{a numeric value for the number of Monte Carlo draws to take when evaluating potential designs}

\item{extra.arguments}{a list of any extra parameters which can be used to control the behaviour of this function or the utility function}

\item{new.ssn.path}{a character string specifying the ssn folder in which to store the output}

\item{ssn}{an object of class SpatialStreamNetwork}

\item{glmssn}{an object of class glmssn}

\item{afv.column}{the name of the column in the SpatialStreamNetwork object that contains the additive function values}

\item{n.points}{the number of points to be included in the final design. This can be a single number, in which case all networks will have the same number of points. This can also be a vector with the same length as the number of networks in ssn.}

\item{utility.function}{a function with the signature utility.function. Users may define their own. This package provides several built-in utility functions, including \code{\link{DOptimality}}, \code{\link{KOptimality}}, and \code{\link{CPOptimality}}.}

\item{prior.parameters}{a function to act as a prior for covariance parameter values}

\item{n.draws}{a numeric value for the number of Monte Carlo draws to take when evaluating potential designs}

\item{extra.arguments}{a list of any extra parameters which can be used to control the behaviour of this function or the utility function}
}
\value{
An object of class SpatialStreamNetwork. The SSNPoints for the obspoints slot will be updated to reflect the selected design.

An object of class SpatialStreamNetwork. The SSNPoints for the obspoints slot will be updated to reflect the selected design.
}
\description{
Finds an optimal design consisting of a specified number of points on a SpatialStreamNetwork.

Finds an optimal design consisting of a specified number of points on a SpatialStreamNetwork.
}
\details{
This function implements the greedy exchange algorithm per Falk et al. (2014) to quickly find an optimal design from a set of sites. The algorithm does not always discover the absolute best design; however, even when this occurs, the algorithm will quickly discover a reasonably effective design.

This function implements the greedy exchange algorithm per Falk et al. (2014) to quickly find an optimal design from a set of sites. The algorithm does not always discover the absolute best design; however, even when this occurs, the algorithm will quickly discover a reasonably effective design.
}
