% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first.loc.R
\name{first.loc}
\alias{first.loc}
\title{Find the first sampling location in a clustered sample}
\usage{
first.loc(ssn.DF.now, start.point.method)
}
\arguments{
\item{ssn.DF.now}{an object of class data.frame which is a single reach's subset of the point.data slot fomr a SpatialStreamNetwork}

\item{start.point.method}{a character vector indicating which method to apply in selecting the the location, with choices of random, prop.to.end, and prop.to.start}
}
\value{
a numeric scalar which is the selected pid column from the point.data data frame
}
\description{
For a sampling design that includes clusters of sampled locations, this function provides the first sample in the cluster.
}
