% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBIDtables.R
\name{getBIDtables}
\alias{getBIDtables}
\title{A function to retrieve the BinaryID tables associated with a SpatialStreamNetwork}
\usage{
getBIDtables(ssn.object, networks = 1:nnetwork(ssn.object))
}
\arguments{
\item{ssn.object}{An object of class SpatialStreamNetwork.}

\item{networks}{A numeric vector of the networks in a SpatialStreamNetwork object for which the binary ID tables should be extracted.}
}
\value{
A named list of length \code{length(networks)} containing the binary ID tables for each of these networks.
}
\description{
Retrieves the Binary ID tables which are stored in a database in the path slot of a SpatialStreamNetwork object and returns them as a list of data.frame objects.
}
\details{
The ssn.object must have a valid folder path stored in its path slot. This folder must contain a binaryID.db database.
}
\examples{

# Single network
s <- createSSN(50, binomialDesign(20), path = tempPath("s.ssn"), importToR = TRUE)
tabs <- getBIDtables(s)

# Multiple networks
s <- createSSN(c(50, 50), binomialDesign(c(20, 20)), path = tempPath("s.ssn"), importToR = TRUE)
tabs <- getBIDtables(s)

}
