% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCalculatedCovMatrix.R
\name{getCalculatedCovMatrix}
\alias{getCalculatedCovMatrix}
\title{A convenience function that extracts the standard errors for the covariance parameters from a \code{glmssn} object}
\usage{
getCalculatedCovMatrix(glmssn, log.scale = FALSE)
}
\arguments{
\item{glmssn}{A fitted \code{glmssn} object.}

\item{log.scale}{Whether the matrix should be on the log-scale. Defaults to \code{FALSE}.}
}
\value{
A covariance matrix for the \code{glmssn} object's covariance parameters.
}
\description{
This function returns the estimated covariance matrix on the covariance parameters from a \code{glmssn} object. Importantly it can be set the return the estimated covariance matrix on its native scale, as opposed to the log scale. This is an extended version of \code{\link{getCalculatedSE}}.
}
\details{
The transformation of the covariance matrix to and from the log-scale is performed using the delta method.
}
