% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCovMat.R
\name{getCovMat}
\alias{getCovMat}
\title{Extract out the covariance matrix for a SpatialStreamNetwork for an arbitrary covariance structure}
\usage{
getCovMat(ssn, glmssn, addfunccol, ...)
}
\arguments{
\item{ssn}{An object of class SpatialStreamNetwork}

\item{glmssn}{A fitted glmssn object. Optional if other arguments are provided.}

\item{addfunccol}{The name of the column in the point.data slot of the obspoints slot of a SpatialStreamNetwork object containing the additive function values used to weight the tail up component of a covariance function.}

\item{...}{Additional arguments to \code{SSN:::makeCovMat}.}
}
\value{
A list containing the covariance matrix and its determinant.
}
\description{
It is sometimes useful to extract out the covariance matrix associated with a covariance structure to test whether it is viable for the data. This function simplifies the process of doing so.
}
