% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExpectedCovMatrix.R
\name{getExpectedCovMatrix}
\alias{getExpectedCovMatrix}
\title{A convenience function that wraps \code{extractExpectedFisherInformation}}
\usage{
getExpectedCovMatrix(glmssn, log.scale = FALSE, ...)
}
\arguments{
\item{glmssn}{A fitted \code{glmssn} object.}

\item{log.scale}{Whether the covariance matrix should be on the log-scale. Defaults to FALSE.}

\item{...}{Optional. Additional arguments to \code{extractExpectedFisherInformation}.}
}
\value{
A vector of expected standard errors for the glmssn object's covariance parameters.
}
\description{
This function returns the expected covariance matrix for the covariance parameters in a \code{glmssn} object. This function is similar to \code{\link{getExpectedSE}}.
}
