% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExpectedSE.R
\name{getExpectedSE}
\alias{getExpectedSE}
\title{A convenience function that wraps \code{extractExpectedFisherInformation}}
\usage{
getExpectedSE(glmssn, log.scale = FALSE, ...)
}
\arguments{
\item{glmssn}{A fitted \code{glmssn} object.}

\item{log.scale}{Whether the standard errors should be given on the log scale. Defaults to FALSE.}

\item{...}{Optional. Additional arguments to \code{extractExpectedFisherInformation}.}
}
\value{
A vector of expected standard errors for the \code{glmssn} object's covariance parameters.
}
\description{
This function returns the expected standard errors on the covariance parameters for a \code{glmssn} object. This function is a more limited version of \code{\link{getExpectedCovMatrix}}. The standard errors are the square roots of the diagonals of the expected information matrix.
}
