% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getImportantMatrices.obs.R
\name{getImportantMatrices.obs}
\alias{getImportantMatrices.obs}
\title{Get the matrices required to calculate the covariance matrix on the data}
\usage{
\code{getImportantMatrices(d.junc, afv = NULL)}
}
\arguments{
\item{d.junc}{An asymmetric, square distance matrix on the design points.}

\item{afv}{The additive function values for each of the design points, in order of their pid.}
}
\value{
A list of 5 matrices in the order dist.hydro, a.mat, b.mat, conn.mat, and w.mat.
}
\description{
Calculates the distance and related matrices between observed or prediction sites on a SpatialStreamNetwork.
}
\details{
This function is called internally by \code{\link{findOptimalDesign}} and \code{\link{doAdaptiveDesign}}. This is done to create the distance and related matrices required to calculate the covariances between observed sites, which is needed for all the utility functions. Note that, although the name of the function suggests this should only work for observed sites, it also works for prediction sites. Therefore, it is also called by \code{\link{findOptimalDesign}} and \code{\link{doAdaptiveDesign}} to create the distance and related matrices among the prediction sites.
}
