% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getImportantMatrices.pxo.R
\name{getImportantMatrices.pxo}
\alias{getImportantMatrices.pxo}
\title{Get the matrices required to calculate the covariance matrix on the data}
\usage{
\code{getImportantMatrices(d.junc.a, d.junc.b, afv.obs = NULL, afv.prd = NULL)}
}
\arguments{
\item{d.junc.a}{A rectangular distance matrix with the observed sites in the rows and prediction sites in the columns.}

\item{d.junc.b}{The transpose matrix of the above.}

\item{afv.obs}{The additive function values for each of the observed points, in order of their pid.}

\item{afv.prd}{The additive function values for each of the prediction points, in order of their pid.}
}
\value{
A list of 5 matrices in the order dist.hydro, a.mat, b.mat, conn.mat, and w.mat.
}
\description{
Calculates the distance and related matrices between observed and prediction sites on a SpatialStreamNetwork.
}
\details{
This is the function called internally by \code{\link{findOptimalDesign}} and \code{\link{doAdaptiveDesign}} to create the matrices required for the evaluation of utility functions that involve the prediction sites.
}
