% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLikelihoodProfile.R
\name{getLikelihoodProfile}
\alias{getLikelihoodProfile}
\title{Extract a likelihood profile that can be plotted}
\usage{
getLikelihoodProfile(glmssn, which.vary, grid.parameters,
  parallelism = "none", n.cores = 1, ...)
}
\arguments{
\item{glmssn}{An object of class \code{glmssn}}

\item{which.vary}{A vector indicating which of the covariance parameters should be varied. You must vary at least one parameter but you can vary all of them if it suits you.}

\item{grid.parameters}{An OPTIONAL argument. You can pass a list giving the grid of parameter values for which to evaluate the log-likelihood surface. This can be left missing. In this case, the function attempts to guess the grid range and resolution based on the log-scale confidence-interval on each of the varying parameters.}

\item{parallelism}{An OPTIONAL argument. This argument indicates what operating system a parallel cluster should be set up for. This option can be one of "none" (for no parallelism), "osx/linux" or "windows". This is not case sensitive. The input should match the user's operating system. Note that this argument is ignored if the next argument \code{n.cores = 1}.}

\item{n.cores}{An OPTIONAL argument. This argument specifies how many CPUs should be assigned to compute the log-likeihoods in parallel. This argument can be any positive integer.}

\item{...}{Any additional arguments to \code{foreach}. Users are NOT recommended to fiddle with this.}
}
\value{
A \code{data.frame} with one column for each covariance parameter and another column for the log-likelihood.
}
\description{
This function computes the likelihood surface in the vicinity of the maximum likelihood estimates of the covariance parameters.
}
\author{
Alan R. Pearse and Jay Ver Hoef
}
