% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStreamDistMatsInOrder.predsxobs.R
\name{getStreamDistMatInOrder.predsxobs}
\alias{getStreamDistMatInOrder.predsxobs}
\title{Retrieve stream distance matrices by network in numeric order}
\usage{
\code{getStreamDistMatInOrder.predsxobs(x)}
}
\arguments{
\item{x}{An object of class SpatialStreamNetwork. The folder where the data for the SpatialStreamNetwork are stored must contain the distance matrix files generated by \code{createDistMat} from the package SSN.}
}
\value{
A list of the distance matrices for the SpatialStreamNetwork, ordered numerically by the network ID.
}
\description{
This function retrieves the distance matrices that have been created for a SpatialStreamNetwork and orders them by network name.
}
\details{
This function addresses the problem that the function \code{getStreamDistMat} returns a list of distance matrices but not in order of network ID. It orders the network IDs in alphanumeric order, meaning that, if a SpatialStreamNetwork has 10 networks, the order will appear as 1, 10, 2, 3, 4, ..., etc.This function differs from \code{\link{getStreamDistMatsInOrder}} in that this function returns only the distance matrices between observed and predicted sites. The other function returns only the distance matrices among the observed and prediction sites, separately.
}
