% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStreamDistMatsInOrder.R
\name{getStreamDistMatInOrder}
\alias{getStreamDistMatInOrder}
\title{Retrieve stream distance matrices by network in numeric order}
\usage{
\code{getStreamDistMatInOrder(x, Name = "obs")}
}
\arguments{
\item{x}{An object of class SpatialStreamNetwork. The folder where the data for the SpatialStreamNetwork are stored must contain the distance matrix files generated by \code{createDistMat} from the package SSN.}

\item{Name}{A string indicating whether the distance matrices for the observed or prediction sites should be extracted. This defaults to "obs" for the observed sites. Replacing this with "preds" will yield the matrices for the prediction sites.}
}
\value{
A list of the distance matrices for the SpatialStreamNetwork, ordered numerically by the network ID.
}
\description{
This function retrieves the distance matrices that have been created for a SpatialStreamNetwork and orders them by network name.
}
\details{
This function addresses the problem that the function \code{getStreamDistMat} returns a list of distance matrices but not in order of network ID. It orders the network IDs in alphanumeric order, meaning that, if a SpatialStreamNetwork has 10 networks, the order will appear as 1, 10, 2, 3, 4, ..., etc.
}
