% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmssn_minimal.R
\name{glmssn_minimal}
\alias{glmssn_minimal}
\title{A function for fitting SpatialStreamNetwork models inside utility functions}
\usage{
\code{glmssn_minimal(..., d, a, b, c, w)}
}
\arguments{
\item{d, a, b, c, w}{Distance matrices (d, a, b), a connectivity matrix (c), and a weights matrix (w).}

\item{...}{Arguments for the function \code{glmssn} from the package SSN.}

\item{nz}{A binary matrix which is 1 when the rows and columns belong to points on the same network, and 0 otherwise.}
}
\value{
An object of class glmssn, as described for the function \code{glmssn} in the package SSN.
}
\description{
This function is the equivalent of glmssn from the package SSN. However, unlike that function, this function requires the user to input the distance and weights matrices for the observed sites over which the model is being fit.
}
\details{
This function is primarily intended for use inside user-defined utility functions, when the utility function requires that models be fit to the data for the design points under consideration. The way this function is set up avoids doubling up on the computations required to generate the distance matrices (which will already have been created prior to the utility function being called) and errors associated with attempting to read in the distance matrices for singular networks.
}
