% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hardCoreDesign.R
\name{hardCoreDesign}
\alias{hardCoreDesign}
\title{The hardCoreDesign function from SSN updated to be compatible with functions from SSNDesign}
\usage{
\code{hardCoreDesign(...)}
}
\arguments{
\item{...}{Arguments for the function \code{hardCoreDesign} in the package SSN.}
}
\value{
An object of class data.frame
}
\description{
\code{hardCoreDesign} replaces a function of the same name from the package SSN.
}
\details{
This function was written to deal with errors resulting in the \code{hardCoreDesign} function from the package SSN when it was used with SpatialStreamNetworks built from real spatial data. It is back-compatible with the \code{createSSN} function from SSN.
}
