% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyObsxPredsABMatrices.R
\name{identifyObsxPredsABMatrices}
\alias{identifyObsxPredsABMatrices}
\title{Identify what elements in a list of distance matrices are for distances between observed and predicted sites.}
\usage{
\code{identifyObsxPredsABMatrices(dist.matrices)}
}
\arguments{
\item{dist.matrices}{A list of distance matrices generated by calling \code{\link{getStreamDistMatsInOrder}} for the prediction sites in a SpatialStreamNetwork.}
}
\value{
A logical vector indicating the positions of the distance matrices between observed and predicted sites.
}
\description{
This function identifies elements in a list of distance matrices which report distances between observed and predicted sites.
}
\details{
This function is called internally inside both \code{\link{getStreamDistMatsInOrder}} and \code{\link{getStreamDistMatsInOrder.predsxobs}}. It is used to identify which elements of a list of distance matrices should be kept or dropped before formatting the distance matrices.
}
