% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importStreams.R
\name{importStreams}
\alias{importStreams}
\title{Import a shapefile of stream edges without observed or predicted sites as a SpatialStreamNetwork}
\usage{
importStreams(filepath, o.write = FALSE)
}
\arguments{
\item{filepath}{a path to a .ssn folder}

\item{o.write}{a logical value}
}
\value{
An object of class SpatialStreamNetwork
}
\description{
\code{importStreams()} imports a SpatialStreamNetwork that has no observed or predicted sites.
}
\examples{
\dontrun{NONE AS YET}

}
