% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locatePointOnEdge.R
\name{locatePointOnEdge}
\alias{locatePointOnEdge}
\title{Place a point at some distance along a stream edge}
\usage{
This function is not intended for direct use. It is called inside \code{\link{generateSites}} to ensure that the placement of simulated sites in geographical space is correct.
}
\arguments{
\item{ssn}{an object of class SpatialStreamNetwork}

\item{rid}{the unique identifier for an edge in a SpatialStreamNetwork}

\item{ratio}{the proportion of the total length of the stream edge where the point should be located}
}
\value{
a vector containing the xy coordinates of the point, and its distance from the first point on the stream edge
}
\description{
\code{locatePointOnEdge()} returns the coordinates of a point along a line, as well as the distance from the first vertex to the point along the line.
}
