% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchBIDridToRownames.R
\name{matchIndices}
\alias{matchIndices}
\title{Get row or element numbers for matches between two vectors}
\usage{
\code{matchIndices} is a function that returns row or element numbers where a set of values from one object occur in another. 
\code{matchIndices(match.from, match.to)}. 

Note, this function is only intended to be used internally by \code{generateSites}.
}
\arguments{
\item{match.from}{a numeric or character vector}

\item{match.to}{a numeric or character vector}
}
\value{
A numeric vector containing the row indices of every match, in the order that they appear in match.from
}
\description{
Get the positions in a vector corresponding to a match between vectors.
}
