% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndpoints.R
\name{ndpoints}
\alias{ndpoints}
\title{Calculate the number of design points on a network}
\usage{
\code{ndpoints(ssn)}
}
\arguments{
\item{ssn}{An object of class SpatialStreamNetwork}
}
\value{
A list with two elements: a vector of design points by network, and a scalar being the total number of design points in the SpatialStreamNetwork object.
}
\description{
\code{ndpoints()} is a function that returns the number of design points on a SpatialStreamNetwork object, both in total and by network.
}
\details{
To find the number of prediction points on a SpatialStreamNetwork object, use \code{\link{nppoints}}.
}
