% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnetwork.R
\name{nnetwork}
\alias{nnetwork}
\title{Find the number of distinct networks in a SpatialStreamNetwork}
\usage{
nnetwork(ssn)
}
\arguments{
\item{ssn}{an object of class SpatialStreamNetwork}
}
\value{
a numeric scalar
}
\description{
Calculates the number of networks are present inside of a SpatialStreamNetwork object.
}
\examples{

s <- createSSN(c(10, 10), binomialDesign(c(2, 2)), path = paste(tempdir(), "tmp.ssn", sep = "/"), importToR = TRUE)
nnetwork(s) # reports 2
     
}
