% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneSampPerSeg.R
\name{OneSampPerSeg}
\alias{OneSampPerSeg}
\title{For a given list of rids, this function randomly selects a single sampling location from each reach.}
\usage{
OneSampPerSeg(ssn.obj, segment.vector)
}
\arguments{
\item{ssn.obj}{an object of class SpatialStreamNetwork}

\item{segment.vector}{a numeric vector of rids for which samples are desired.}
}
\value{
a numeric vector of pids
}
\description{
For a given list of rids, this function randomly selects a single sampling location from each reach.
}
