% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overwriteBeta.R
\name{overwriteBeta}
\alias{overwriteBeta}
\title{Overwrite values of the fixed effects parameters in a fitted glmssn object.}
\usage{
overwriteBeta(glmssn, new.beta)
}
\arguments{
\item{glmssn}{An object of class glmssn.}

\item{new.beta}{A numeric vector containing the new values of the fixed effects parameters.}
}
\value{
An obejct of class glmssn with the old values of theta replaced by the new ones.
}
\description{
It is not recommended to manually assign values for the fixed effects parameters in a glmssn object, since the best estimates are already provided. However, it may be necessary for data simulation purposes.
}
\details{
Note that the length of new.beta must be equal to the length of \code{glmssn$estimates$betahat} and the fixed effects parameter values must be provided in the same order.
}
