% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overwriteThetas.R
\name{overwriteThetas}
\alias{overwriteThetas}
\title{Overwrite values of the covariance parameters in a fitted glmssn object.}
\usage{
overwriteThetas(glmssn, new.theta)
}
\arguments{
\item{glmssn}{An object of class glmssn.}

\item{new.theta}{A numeric vector containing the new values of the covariance parameters.}
}
\value{
An obejct of class glmssn with the old values of theta replaced by the new ones.
}
\description{
It is not recommended to manually assign values for the covariance parameters in a glmssn object, since the best estimates are already provided. However, it may be necessary for data simulation purposes.
}
\details{
Note that the length of new.theta must be equal to the length of \code{glmssn$estimates$theta} and the covariance parameter values must be provided in the same order.
}
