% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-method.R
\name{plot.ssndesign}
\alias{plot.ssndesign}
\title{Plot trace of the Greedy Exchange Algorithm}
\usage{
\method{plot}{ssndesign}(x, y,
  which.iteration = 1:length(x$trace.per.random.start), legend = TRUE,
  ...)
}
\arguments{
\item{x}{An object of class \code{ssndesign}}

\item{y}{Not used}

\item{which.iteration}{A numeric indicating the trace that should be plotted. This can be a vector, in which case all traces corresponding to those indices will be plotted. All traces are plotted by default.}

\item{legend}{Whether the legend should be drawn on the plot. Defaults to \code{FALSE}. Not advised if there are many traces because the legend will be too large to properly fit on the plot.}

\item{...}{Additional arguments to the \code{plot} function.}
}
\description{
A generic plot function adapted for `ssndesign` objects. It plots the traces from the optimisation algorithm.
}
\details{
The plot method produces line plots tracing the expected utility of the best design in the algorithm for each set of coordinate exchanges.
}
