% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrace.R
\name{plotTrace}
\alias{plotTrace}
\title{Plot the diagnostics output by the design functions}
\usage{
\code{plotTrace(ssndesign)}
}
\arguments{
\item{ssndesign}{The list object output by one of the main design functions.}
}
\value{
A ggplot object
}
\description{
Creates a traceplot of the maximum utilities
}
\details{
This function plots the trace for the utility function as the main design functions iterate through the greedy exchange algorithm. 

Note that this function simply makes use of the output from ssndesign2DF in a standard format. Users can simply use \code{ssndesign2DF} to obtain the same data for plotting, which they can use in whatever fashion they wish.
}
