% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_ind.R
\name{rs_ind}
\alias{rs_ind}
\title{Random Sampling from a Spatial Stream Network}
\usage{
rs_ind(ssn = ssn.obj, size = 20)
}
\arguments{
\item{ssn}{an object of \code{\link[SSN]{SpatialStreamNetwork-class}} the observations of which are to be subsetted by random sampling}

\item{size}{a number. the size of the random sample of observations we will draw from SSN}
}
\description{
rs_ind generates an index vector for random sampling of observations from a \code{\link[SSN]{SpatialStreamNetwork-class}} object with the the function \code{\link[SSN]{subsetSSN}}.
}
\examples{
# simulate an SSN object from which to sample
ssn.1 <- createSSN(n = 1e2, obsDesign = systematicDesign(spacing = 0.5), path = paste0(tempdir(),'/sim_ssn_',sample(x = 1:100, size = 1),'1.ssn'), importToR = TRUE)
# generate the logical vector we will use to subset the SSN object
index.rnd <- rs_ind(ssn = ssn.1, size = 20)
rs1.ssn.1 <- subsetSSN(ssn = ssn.1,
              filename = paste0(tempdir(),'/random_sample',sample(x = 1:100, size = 1),'.ssn'),
              subset = index.rnd)
# plot the original set of observations on the network
plot(ssn.1, main = 'the original set of observations on the SSN')
# plot the random sample from the original set of observations on the network
dev.new()
plot(rs1.ssn.1, main = 'a random sample of 20 observations from\\nthe original set of observations on the SSN')
}
