% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SameSegmentClusters.sample.R
\name{SameSegmentClusters.sample}
\alias{SameSegmentClusters.sample}
\title{For sampling designs including clusters of samples, this function selects clustered samples within each reach that is provided.}
\usage{
\code{SameSegmentClusters.sample(ssn.obj, ClustDistMethod="prop.shortest.seg", segment.vector, max.dist=NA, cluster.size, start.point.method="random", bin.table)}
}
\arguments{
\item{ssn.obj}{an object of class SpatialStreamNetwork}

\item{ClustDistMethod}{a character vector providing the method for how clustered samples are selected from within a single reach.}

\item{segment.vector}{a numeric vector of rids for which clustered samples are desired.}

\item{max.dist}{an optional distance provided as the upper limit for a distance between two locations within a common reach whereby those locations can be selected into a common cluster.}

\item{start.point.method}{a character vector providing the method for how the first location in each cluster should be selected}

\item{bin.table}{an object of class data.frame which represents the bindaryID object for a SpatialStreamNetwork}
}
\value{
a data.frame
}
\description{
For sampling designs including clusters of samples, this function selects clustered samples within each reach that is provided.
}
