% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchAllDesigns.R
\name{searchAllDesigns}
\alias{searchAllDesigns}
\title{Search all possible designs to find one which maximises the utility}
\usage{
\code{searchAllDesigns(ssn, glmssn, afv.column, n.points, utility.function, prior.parameters, n.draws = 500, extra.arguments = NULL)}
}
\arguments{
\item{ssn}{an object of class SpatialStreamNetwork}

\item{n.points}{the number of points to be included in the final design. This can be a single number, in which case all networks will have the same number of points. This can also be a vector with the same length as the number of networks in ssn.}

\item{utility.function}{a function with the signature utility.function. Users may define their own. This package provides several built-in utility functions, including \code{\link{sequentialDOptimality}}, \code{\link{KOptimality}}, and \code{\link{sequentialCPOptimality}}.}

\item{prior.parameters}{a function to act as a prior for covariance parameter values}

\item{n.draws}{a numeric value for the number of Monte Carlo draws to take when evaluating potential designs}

\item{extra.arguments}{a list of any extra parameters which can be used to control the behaviour of this function or the utility function}

\item{glmssn}{an object of class glmssn}

\item{afv.column}{the name of the column in the SpatialStreamNetwork object that contains the additive function values}
}
\value{
An object of class SpatialStreamNetwork. The SSNPoints for the obspoints slot will be updated to reflect the selected design.
}
\description{
This function takes all possible combinations of n design points and evaluates a utility function for each combination.
}
\details{
This is very computationally expensive. Only use this function for small examples or to benchmark the performance of the greedy exchange algorithm.
}
