% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequentialEDOptimality.R
\name{sequentialEDOptimality}
\alias{sequentialEDOptimality}
\title{A utility function for empirical parameter estimation (sequential)}
\usage{
\code{sequentialEDOptimality(ssn, glmssn, design.points, prior.parameters, n.draws, extra.arguments)}
}
\arguments{
\item{ssn}{An object of class SpatialStreamNetwork}

\item{glmssn}{An model object of class glmssn.}

\item{design.points}{A vector of pids corresponding to a set of observed sites in the obspoints slot of the SpatialStreamNetwork object.}

\item{prior.parameters}{A list of random functions that are parameterised in terms of n.draws.}

\item{n.draws}{A numeric scalar for the number of Monte Carlo draws to use when approximating the utility.}

\item{extra.arguments}{A list of extra parameters that control the behaviour of the utility function. The distance matrices required to compute covariance matrices are also stored in this list. Note that these are generated inside \code{\link{findOptimalDesign}} and \code{\link{doAdaptiveDesign}}.}
}
\value{
A numeric scalar.
}
\description{
\code{sequentialEDOptimality} is a utility function that can be used with \code{\link{doAdaptiveDesign}}. It is a utility function that minimises the determinant of the variance-covariance matrix of the fixed effects for models that are fit over a set of design points.
}
\details{
Note that this function operates differently to \code{\link{EDOptimality}}. The functions \code{\link{DOptimality}} and \code{\link{EDOptimality}} assume there are no sites which have already been incorporated into a design. They compute the variance-covariance matrix on the fixed effects (Som et al., 2014) for each set of simulated or esimated covariance parameters, respectively. In this sequential form, the observed variance-covariance matrix is extracted for the sites which are fixed in the design. Then the sites which are not fixed are used to estimate the expected variance-covariance matrix. The observed and expected matrices are added, before being reduced to a scalar value to serve as the utility.
}
