% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateFromSSNM.R
\name{simulateFromSSNM}
\alias{simulateFromSSNM}
\title{Simulate data, with error, from a glmssn object}
\usage{
\code{simulateFromSSNM(ssn, glmssn, fixed.effects = NULL, covariance.type = NULL, covariance.parms = NULL)}
}
\arguments{
\item{ssn}{An object of class SpatialStreamNetwork}

\item{glmssn}{An object of class glmssn}

\item{fixed.effects}{A numeric vector specifying the values of the fixed effects. Leave as NULL unless you specifically intend to override the information about the fixed effects you get from the glmssn.}

\item{covariance.type}{A character vector specifying what kind of covariance structure should be used to simulate the data. Leave as NULL unless you specifically intend to overwrite the covariance components from the glmssn.}

\item{covariance.parms}{A numeric vector specifiying the values of the covariance parameters. Leave as NULL unless you specifically intend to override the informaiton about the covariance parameters you get from the glmssn.}
}
\value{
An object of class SSNM, which is a modified version of the one that was given as the argument ssn.
}
\description{
This function simulates data on a SpatialStreamNetwork based on information from a fitted model object. It essentially acts as a wrapper for SimulateOnSSN.
}
\details{
It is only possible to use this function to simulate data on a SpatialStreamNetwork object when a fitted model exists.
}
