% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateOnSSN_minimal.R
\name{SimulateOnSSN_minimal}
\alias{SimulateOnSSN_minimal}
\title{A function to do simulate on SSN on a single network with pre-defined matrices}
\usage{
\code{SimulateOnSSN_minimal(..., matrices.obs, net.zero.obs, matrices.preds, net.zero.preds = NULL, matrices.predsxobs = NULL, net.zero.predsxobs = NULL)}
}
\arguments{
\item{...}{Arguments for the function \code{SimulateOnSSN} from the package SSN.}

\item{d, a, b, c, w}{Distance matrices (d, a, b), a connectivity matrix (c), and a weights matrix (w).}
}
\value{
An object of class SpatialStreamNetwork, as described for the function \code{SimulateOnSSN} in the package SSN.
}
\description{
This is a modified version of the function \code{SimulateOnSSN} from the package SSN, as written by Jay ver Hoef. This version of the function is identical in every respect, except that it accepts pre-computed distance matrices instead of creating them from scratch inside the function.
}
\details{
This function is primarily intended for use inside user-defined utility functions, when the utility function requires that predictions from models (with error) must be made for the design points under consideration. The way this function is set up avoids doubling up on the computations required to generate the distance matrices (which will already have been created prior to the utility function being called) and errors associated with attempting to read in the distance matrices for singular networks.
}
