% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceFillingMaxiMin.R
\name{spaceFillingMaxiMin}
\alias{spaceFillingMaxiMin}
\title{A utility function for maximin space-filling designs}
\usage{
spaceFillingMaxiMin(ssn, glmssn, design.points, prior.parameters, n.draws,
  extra.arguments)
}
\arguments{
\item{ssn}{Ignored in this function but this argument is present so this function is consistent with all other utility functions.}

\item{glmssn}{Ignored in this function but this argument is present so this function is consistent with all other utility functions.}

\item{design.points}{A vector of pids corresponding to a set of observed sites in the obspoints slot of the SpatialStreamNetwork object.}

\item{prior.parameters}{Ignored in this function but this argument is present so this function is consistent with all other utility functions.}

\item{n.draws}{Ignored in this function but this argument is present so this function is consistent with all other utility functions.}

\item{extra.arguments}{A list of extra parameters that control the behaviour of the utility function. The distance matrices required to compute covariance matrices are also stored in this list. Note that these are generated inside \code{\link{optimiseSSNDesign}}.}
}
\value{
The expected utility of a design. This number represents the smallest non-zero distance between any pair of sites in the design.
}
\description{
\code{spaceFillingMaximin} is a utility function used by \code{\link{constructSpaceFillingDesign}}. It is also fully compatible with \code{\link{optimiseSSNDesign}}. It is a utility function that maximises the minimum interpoint distance between a set of design points.
}
\details{
\code{spaceFillingMaximin} is deterministic and will ignore the arguments \code{prior.parameters} and \code{n.draws}. It is expected to be very fast, unless there are several hundred design points to select. 

For examples, please see \code{\link{constructSpaceFillingDesign}}.
}
