% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spliceSSNSites.R
\name{spliceSSNSites}
\alias{spliceSSNSites}
\title{Introduce new sites into a SpatialStreamNetwork object from a shapefile}
\usage{
\code{spliceSSNSites(ssn, splice.obs, splice.preds = NULL)}
}
\arguments{
\item{ssn}{An object of class SpatialStreamNetwork}

\item{new.ssn.path}{A file path to a folder to store the augmented ssn.}

\item{splice.obs}{A file path for the shapefile of observed sites which should be brought into the ssn object.}

\item{splice.preds}{Optionally a file path for a shapefile of prediction points which should also be brought into the ssn object. This can be (and is by default) NULL if there are no such prediction points.}
}
\value{
An object of class SpatialStreamNetwork that contains the old and new sites.
}
\description{
This function does the reverse of \code{\link{splitSSNSites}}. Instead of separating a \code{SpatialStreamNetwork} object into different shapefiles, this function combines a set of observed points from a shapefile with an existing SSN object.
}
\details{
As for \code{splitSSNSites}, this function is wrapped by \code{\link{doAdaptiveDesign}}. While \code{doAdaptiveDesign} should be sufficient for solving most adaptive design problems, more specialised or unusual applicaitons may require this function to be used separately.
}
\examples{

\dontrun{#code}
 
}
