% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitSSNSites.R
\name{splitSSNSites}
\alias{splitSSNSites}
\title{Split the temporally replicated point data for a SpatialStreamNetwork object into separate shapefile}
\usage{
splitSSNSites(big.ssn, new.ssn.path, replication.variable, preds = TRUE,
  preferred.wd = getwd())
}
\arguments{
\item{big.ssn}{The SpatialStreamNetwork object which contains the temporally replicated site data.}

\item{new.ssn.path}{A file path for a new SSN object.}

\item{replication.variable}{A string which indicates which variable in the \code{big.ssn@obspoints@SSNPoints[[1]]@point.data} should be used to split big.ssn into separate shapefiles.}

\item{preds}{A logical which indicates whether the prediction sites should be dealt with in the same way. This predpoints object should contain the same levels of the replication variable as the obspoints slot.}

\item{preferred.wd}{A directory where the shapefiles should be written.}
}
\value{
This function returns an object of class SpatialStreamNetwork. This SpatialStreamNetwork will contain the sites with the lowest value of the replication variable.
}
\description{
Primarily intended to be used with the \code{\link{optimiseSSNDesign}} function, this function takes an SSN with temporal replicates across some number of sites and splits the sites into separate shapefiles based on the levels of the replication variable. This makes it easier to deal with the SSN in a statistically consistent way when performing sequential design over the network.
}
\details{
This function is wrapped by \code{\link{optimiseSSNDesign}}. The \code{optimiseSSNDesign} function should be sufficient for most adaptive design problems where this is required. However, for more bespoke applications, it may be worthwhile to use this function on its own instead.
}
\examples{

\dontrun{

s <- createSSN(10, binomialDesign(100, 2), path = tempPath("r.ssn"), importToR = T)

# Split SSN sites into two shapefiles, one for each year
split <- splitSSNSites(s, tempPath("split.ssn"), "Time", FALSE, tempdir())

# Join the year 2 shapefile back to the year 1 shapefile in the SSN
spliced <- spliceSSNSites(split, tempPath("spliced.ssn"), paste0(tempdir(), "/sites2.shp"))

}

}
