% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRS.sample.R
\name{SRS.sample}
\alias{SRS.sample}
\title{This function selects a simple random sample of points from a stream network object of class SpatialStreamNetwork.}
\usage{
\code{SRS.sample(ssn.obj, sample.size)}
}
\arguments{
\item{ssn.obj}{an object of class SpatialStreamNetwork}

\item{sample.size}{an integer indicating the desired number of point locations selected for the sample.}
}
\value{
a data.frame
}
\description{
This function selects a simple random sample of points from a stream network object of class SpatialStreamNetwork.
}
