% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSNDesign.R
\docType{package}
\name{SSNDesign}
\alias{SSNDesign}
\alias{SSNDesign-package}
\title{SSNDesign: Optimal and adaptive design for Spatial Stream Networks}
\description{
The adssn package builds on the SSN package to provide a framework for solving design problems on SpatialStreamNetwork objects.
}
\section{SSNDesign functions}{

the SSNDesign package contains the following functions:
- constructTotalMatrix
- CPOptimality
- doAdaptiveDesign
- doAdaptiveDesignParallel
- DOptimality
- EDOptimality
- EKOptimality
- findOptimalDesign
- findOptimalDesignParallel
- generateSites
- getBIDtables
- getImportantMatrices.obs
- getImportantMatrices.pxo
- getStreamDistMatsInOrder.predsxobs
- getStreamDistMatsInOrder
- glmssn_minimal
- hardCoreDesign
- importStreams
- KOptimality
- ndpoints
- nnetwork
- nppoints
- searchAllDesigns
- sequentialCPOptimality
- sequentialDOptimality
- sequentialEDOptimality
- SimulateOnSSN_minimal
- stableInverse
- systematicDesign
}

