% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssndesign2DF.R
\name{ssndesign2DF}
\alias{ssndesign2DF}
\title{Pull the design diagnostics into a dataframe.}
\usage{
\code{ssndesign2DF(ssndesign)}
}
\arguments{
\item{ssndesign}{The list object output by one of the main design functions.}
}
\value{
A data.frame object.
}
\description{
Creates a dataframe of utility values across all networks for all exchange algorithm hyper-iterations from the list output by the main design functions.
}
\details{
This function does stuff.
}
