% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdsGs.R
\name{stdsGs}
\alias{stdsGs}
\title{Transform to standard normal}
\usage{
stdsGs(x, na.rm = T)
}
\arguments{
\item{x}{A numeric vector of observations}

\item{na.rm}{Whether missing observations should be ignored. Defaults to \code{TRUE}.}
}
\value{
A numeric vector of size \code{length(x)}.
}
\description{
This function takes the vector x and transforms it to be approximately standard normal. It is meant to be used with the function \code{\link{transformSSNVars}}.
}
\examples{

x <- rnorm(100, 4, 2)
x.std <- stdsGs(x)
par(mfrow = c(1, 2))
plot(density(x), main = "Unstandardised")
plot(density(x.std), main = "Standardised")

}
