% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stream.Network.Samples.R
\name{Stream.Network.Samples}
\alias{Stream.Network.Samples}
\title{For a Spatial Stream Network with a dense grid of potential sampling locations, this function draws samples according to the sampling designs found most optimal in Som et al. (2014), and a few others.}
\usage{
\code{Stream.Network.Samples(ssn.list, sample.method, sample.size,cluster.number=NULL, Wmat=NULL, DistMat=NULL, ...)}
}
\arguments{
\item{ssn.list}{an object of class list that contains one item names "ssn.object" which is a SpatialStreamNetwork and another item named "bin.table" which is the binaryID table for the assoicated SpatialStreamNetwork}

\item{sample.method}{a character vector providing the label for the specific sampling design that is desired, with references to Som et al. (2014) including simple random sample "SRS", "GRTS", "GRTSmouth", "GRTSclus", H1 "Extreme.Clust.and.Singles", C3 "Trib.Sets.Ext.Singles.sample", C4 "Trib.Sets.Ext.Singles.Mouth.sample"}

\item{sample.size}{a numeric scaler with the desired sample size (USUALLY) but can differ depending on sampling design chosen - something to clean up.}

\item{cluster.number}{a numeric scaler for the number of sampling locations contained within each clusted sample.}

\item{Wmat}{leftover from previous version of function when weigths matrix required as input}

\item{DistMat}{leftover from previous version of function when distance matrix required as input}
}
\value{
a list containing a SpatialStreamNetwork for the selected samples, and a numeric vector with the selected pids
}
\description{
For a Spatial Stream Network with a dense grid of potential sampling locations, this function draws samples according to the sampling designs found most optimal in Som et al. (2014), and a few others.
}
