% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempPath.R
\name{tempPath}
\alias{tempPath}
\title{Generate paths for .ssn folders to be written to tempdir()}
\usage{
tempPath(ssn.name)
}
\arguments{
\item{ssn.name}{The name of the ssn folder (including .ssn extension) as a string.}
}
\value{
The path for the \code{ssn.name} folder in R's temporary directory.
}
\description{
This function takes the name of a .ssn folder and returns the full path to that .ssn folder in R's temporary directory.
}
\examples{

new.path <- tempPath("s.ssn")
s <- createSSN(5, binomialDesign(10), path = new.path, importToR = TRUE)

}
