% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformSSNVars.R
\name{transformSSNVars}
\alias{transformSSNVars}
\title{A function to conveniently transform multiple variables inside the observed and predicted points' data slots in a SpatialStreamNetwork object}
\usage{
transformSSNVars(ssn, new.ssn.path, obs.vars, preds.vars, func,
  write.out = FALSE, ...)
}
\arguments{
\item{ssn}{An object of class SpatialStreamNetwork}

\item{new.ssn.path}{The directory in which a new .ssn directory should be written out. Ignored if write.out = FALSE.}

\item{obs.vars}{A character vector containing the names of the columns in the observed sites point.data slot which should be transformed. If this is a named vector, then new columns with those names will be added. Otherwise, the old columns will be overwritten.}

\item{preds.vars}{A character vector containing the names of the columns in the prediction sites point.data slot which should be transformed. If this is a named vector, then new columns with those names will be added. Otherwise, the old columns will be overwritten.}

\item{func}{The function which should be used to transform the columns.}

\item{write.out}{A logical indicating whether the .ssn folder in \code{ssn@path} should be updated.}

\item{...}{Additional arguments to func.}
}
\value{
An object of class SpatialStreamNetwork
}
\description{
The process of modifying the data.frame objects in the observed and predicted sites of a SpatialStreamNetwork can be complicated. This function simplifies this process.
}
\details{
Note, if write.out = TRUE then the SSN will be re-imported before returning. Note also that write.out = TRUE will cause the data inside the .ssn directory to be overwritten.
}
