% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpackExampleData.R
\name{unpackExampleData}
\alias{unpackExampleData}
\title{Copy package data to a new folder}
\usage{
unpackExampleData(unpack.loc)
}
\arguments{
\item{unpack.loc}{The path to some folder on the user's computer that the files should be copied to.}
}
\value{
A logical indicating whether file transfer was successful.
}
\description{
The SSNdesign package contains several .ssn and .Rdata files. These are the files referred to in the package vignette. This function takes these files out of SSNdesign's folders and copies them to a user-specified location on the user's computer.
}
\details{
Please note that the function will return \code{FALSE}, indicating unsuccessful file transfer, if the directory specified in \code{unpack.loc} already contains files with the same names as the example data files.
}
\examples{

\dontrun{

# To the documents folder:
unpackExampleData("~/")
}

}
