% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateDesigns.R
\name{validateDesigns}
\alias{validateDesigns}
\title{Validate designs on SpatialStreamNetworks}
\usage{
validateDesigns(designs.list, glmssn, type, dist = rmse, n.sims = 500)
}
\arguments{
\item{full.ssn}{An object of class SpatialStreamNetwork from which the designs were originally selected.}

\item{designs.list}{A SpatialStreamNetwork objects containing the design points to be used for valdiation.}

\item{glmssn}{A fitted glmssn object}

\item{type}{A string indicating what the design(s) should be tested on. This argument must be one of "fixed.effects", "covariance.parms" or "prediction".}

\item{n.sims}{The number of times values should be simulated for the validation procedure. Defaults to 500.}

\item{dist}{A loss function comparing the simulated and predicted values at each step in the validation process. See Details for more information.}
}
\value{
A list of validation results. See Details for more information.
}
\description{
This function performs validation on
}
\details{
The functions given to the argument dist should be functions of the form \code{function(x, y)}. They must at least accept two arguments. If the ordering of the arguments is important, then note that the simulated values are passed to the x argument and the predicted values to the y argument. 

The list of validation results returned by this function contains a matrix of dimensions m x n.sims, where m is equal to length(designs.list). The elements of this matrix represent a single validation result for one design at one simulation. It also contains a numeric vector of m elements, which are obtained by taking the rowmeans of the aforementioned matrix.
}
