% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateShreveStreamOrder.R
\name{calculateShreveStreamOrder}
\alias{calculateShreveStreamOrder}
\title{A function to calculate the Shreve stream order of a stream edge}
\usage{
calculateShreveStreamOrder(binary.id.table)
}
\arguments{
\item{binary.id.table}{A data.frame representing a binaryID table extracted from the binaryID.db in the path of a SpatialStreamNetwork.}
}
\value{
A data.frame with two columns: the first is the RID of each stream segment; the second is the Shreve order of each segment.
}
\description{
For a given binaryID table, this function quickly computes the Shreve stream order of every stream segment by counting the number of upstream inlet edges there are.
}
