% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPOptimality.R
\name{CPOptimality}
\alias{CPOptimality}
\title{A utility function for covariance parameter estimation}
\usage{
\code{CPOptimality(ssn, glmssn, design.points, prior.parameters, n.draws, extra.arguments)}
}
\arguments{
\item{ssn}{An object of class SpatialStreamNetwork}

\item{glmssn}{A fitted model object of class glmssn.}

\item{design.points}{A vector of pids corresponding to a set of observed sites in the obspoints slot of the SpatialStreamNetwork object.}

\item{prior.parameters}{A list of random functions that are parameterised in terms of n.draws.}

\item{n.draws}{A numeric scalar for the number of Monte Carlo draws to use when approximating the utility.}

\item{extra.arguments}{A list of extra parameters that control the behaviour of the utility function. The distance matrices required to compute covariance matrices are also stored in this list. Note that these are generated inside \code{\link{optimiseSSNDesign}}.}
}
\value{
A single number representing the expected utility for the design specified by \code{design.points}.
}
\description{
The function \code{CPOptimality} implements the CP-optimal utility function from Falk et al. (2014).
}
\details{
The utility function is

\deqn{U(d,\theta) = \log \det \left(I(\theta)\right)}{U(d, \theta) = log(det(I(\theta)))}

where \eqn{I(\theta)} is the expected Fisher information matrix for the covariance parameters \eqn{\theta}.

A value of \code{-1e9} indicates a failure of this utility function.

One final note: do not worry about passing arguments to this function. All arguments are dealt with internally by \code{\link{optimiseSSNDesign}}.
}
\examples{

\dontrun{
# Create stream network
s <- createSSN(100, systematicDesign(0.25), path = paste(tempdir(), "s.ssn", sep = "/"), importToR = TRUE)
createDistMat(s)

# Simulate data on network
s <- SimulateOnSSN(s, getSSNdata.frame(s),formula = ~1, coefficients = 1, CorParms = c(1,2,1,2,1,2,0.1),addfunccol = "addfunccol")$ssn.object

# Fit a model to the simulated data
m <- glmssn(Sim_Values ~ 1, s, addfunccol = "addfunccol")

# Define the priors on the covariance parameters
p <- constructLogNormalCovPriors(m)

# Find the optimal design using D-optimality as the utility function
r <- optimiseSSNDesign(s, paste(tempdir(), "r.ssn", sep = "/"), m, 25, utility.function = CPOptimality, prior.parameters = p)

# Plot result to check
plot(r$ssn.new, "Sim_Values")
}

}
