% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDOptimality.R
\name{EDOptimality}
\alias{EDOptimality}
\title{A utility function for optimal designs for the empirical estimation of fixed effects parameters.}
\usage{
\code{EDOptimality(ssn, glmssn, design.points, prior.parameters, n.draws, extra.arguments)}
}
\arguments{
\item{ssn}{An object of class SpatialStreamNetwork}

\item{glmssn}{A fitted model object of class glmssn.}

\item{design.points}{A vector of pids corresponding to a set of observed sites in the obspoints slot of the SpatialStreamNetwork object.}

\item{prior.parameters}{A list of random functions that are parameterised in terms of n.draws.}

\item{n.draws}{A numeric scalar for the number of Monte Carlo draws to use when approximating the utility.}

\item{extra.arguments}{A list of extra parameters that control the behaviour of the utility function. The distance matrices required to compute covariance matrices are also stored in this list. Note that these are generated inside \code{\link{optimiseSSNDesign}}.}
}
\value{
A single number representing the expected utility for the design specified by \code{design.points}.
}
\description{
The function \code{EDOoptimality} implements the ED-optimal utility function from Falk et al. (2014).
}
\details{
The utility function is

\deqn{U(d, \theta, y) = \log\det(\hat{Var}(\beta)^{-1})}{U(d, \theta, y) = log(det(\hat{Var}(\beta)^{-1}))}

where \eqn{\hat{Var}(\beta)} is the estimated variance-covariance matrix of the fixed effects from a fitted \code{glmssn} object.

One final note: do not worry about passing arguments to this function. All arguments are handled internally by \code{\link{optimiseSSNDesign}}.
}
\examples{

\dontrun{

# Create stream network
s <- createSSN(100, systematicDesign(0.25), path = paste(tempdir(), "s.ssn", sep = "/"), importToR = TRUE)
createDistMat(s)

# Simulate data on network
s <- SimulateOnSSN(s, getSSNdata.frame(s), formula = ~1, coefficients = 1, CorParms = c(1,2,1,2,1,2,0.1),addfunccol = "addfunccol")$ssn.object

# Fit a model to the simulated data
m <- glmssn(Sim_Values ~ 1, s, addfunccol = "addfunccol")

# Define the priors on the covariance parameters
p <- constructLogNormalCovPriors(m)

# Find the optimal design using D-optimality as the utility function
r <- optimiseSSNDesign(s, paste(tempdir(), "r.ssn", sep = "/"), m, 25, utility.function = EDOptimality, prior.parameters = p)

# Plot result to check
plot(r$ssn.new, "Sim_Values")

}

}
\references{
Falk, M., McGree, J.M., and Pettit, A.N. (2014). Sampling designs on stream networks using the pseudo-Bayesian approach. \emph{Environmental and Ecological Statistics}, \emph{21}(4), 751-773.
}
