% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateFixedDesigns.R
\name{evaluateFixedDesigns}
\alias{evaluateFixedDesigns}
\title{Evaluate fixed designs}
\usage{
\code{evaluateFixedDesigns(ssn, glmssn, afv.column, list.designs, utility.function, prior.parameters, n.draws = 500, extra.arguments = NULL)}
}
\arguments{
\item{ssn}{an object of class SpatialStreamNetwork}

\item{glmssn}{an object of class glmssn}

\item{afv.column}{the name of the column in the SpatialStreamNetwork object that contains the additive function values}

\item{list.designs}{a list containing vectors of}

\item{utility.function}{a function with the signature utility.function. Users may define their own. This package provides several built-in utility functions.}

\item{prior.parameters}{a function to act as a prior for covariance parameter values}

\item{n.draws}{a numeric value for the number of Monte Carlo draws to take when evaluating potential designs}

\item{extra.arguments}{a list of any extra parameters which can be used to control the behaviour of this function or the utility function}
}
\value{
A dataframe of design ID (integers from 1 in the same order as list.designs); Size, the number of design points; and, Utility, the utility value.
}
\description{
Evaluates the utility of a list of user-specified designs.
}
\details{
This function takes a list of user-specified designs for a spatial stream network and evaluates some utility function over those designs. This function should be relatively fast compared to the functions which search for an optimal design. However, because this function is not parallellised, it is recommended that users limit themselves to fewer than one hundred designs. Note, this function is called internally in the design diagnostic functions.
}
