% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSites.R
\name{generateSites}
\alias{generateSites}
\title{Simulate observed and/or predicted points on an empty SpatialStreamNetwork.}
\usage{
generateSites(ssn, edgeweights, obsDesign, predDesign = noPoints,
  importToR = FALSE, treeFunction = igraphKamadaKawai)
}
\arguments{
\item{ssn}{an object of class SpatialStreamNetwork}

\item{edgeweights}{a string; the name of a column in the data slot of the SpatialStreamNetwork that is used to weight edges}

\item{obsDesign}{a design function}

\item{predDesign}{a design function}

\item{importToR}{a logical value}

\item{treeFunction}{DEPRECATED}
}
\value{
An object of class SpatialStreamNetwork
}
\description{
\code{generateSites()} simulates observed and predicted sites on an empty SpatialStreamNetwork
}
\section{Warning}{

This function is currently written such that any observed or predicted sites generated on a SpatialStreamNetwork will overwrite any existing sites.
}

\examples{
\dontrun{NONE AS YET}

}
