% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSites.R
\name{generateSites}
\alias{generateSites}
\title{Simulate observed and/or predicted points on an empty SpatialStreamNetwork.}
\usage{
generateSites(ssn, obsDesign, predDesign = SSN:::noPoints,
  o.write = FALSE)
}
\arguments{
\item{ssn}{an object of class SpatialStreamNetwork}

\item{obsDesign}{a design function from the package SSN.}

\item{predDesign}{a design function from the package SSN. By default, no prediction points are generated using the function \code{noPoints}.}

\item{o.write}{a logical indicating whether any existing observed and/or predicted sites in \code{ssn@path} should be overwritten. Defaults to \code{FALSE}.}
}
\value{
An object of class SpatialStreamNetwork
}
\description{
This function takes an empty SpatialStreamNetwork and creates a set of proposed design points on the stream network.
}
\details{
This function works in a similar fashion to \code{createSSN} from the package SSN. This function uses one of the design functions provided in SSN to generate a set of sites on an empty SpatialStreamNetwork. The differences with this function are that (1) this function assumes the stream network already exists and concentrates only on generating the site locations and associated data, and (2) locates the sites in physical space with physical coordinates. The function \code{createSSN} only locates sites on an abstract network which is not associated with a coordinate reference system, hence limiting its applications to real data.

The new observed and/or predicted sites will appear as shapefiles in \code{ssn@path}.
}
\examples{

# SSN with small random design
s <- createSSN(100, binomialDesign(20), 
path = tempPath("example02.ssn"), importToR = TRUE)

# Overwrite design with a systematic design
s2 <- generateSites(s, systematicDesign(0.5), o.write = TRUE) 

}
