% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importStreams.R
\name{importStreams}
\alias{importStreams}
\title{Import a shapefile of stream edges without observed or predicted sites as a SpatialStreamNetwork}
\usage{
\code{importStreams(filepath, o.write = FALSE)}
}
\arguments{
\item{filepath}{a path to a folder containing a stream network. The folder should have the extension .ssn.}

\item{o.write}{a logical value indicating whether the binaryID.db inside the folder should be overwritten if it already exists. Note that, if no binaryID.db exists, that database will be generated by this function.}
}
\value{
An object of class SpatialStreamNetwork, with nothing in obspoints and predpoints slots.
}
\description{
Imports a SpatialStreamNetwork that has no observed or predicted sites.
}
\details{
The purpose of this function is to allow users to ingest stream shapefiles with no observed or prediction sites into R as a SpatialStreamNetwork. This will then allow them to use their data with functions like \code{generateSites}, which will allow them to explore various designs on their stream network.
}
