% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchBIDridToRownames.R
\name{matchIndices}
\alias{matchIndices}
\title{Get row or element numbers for matches between two vectors}
\usage{
matchIndices(match.from, match.to)
}
\arguments{
\item{match.from}{a numeric or character vector}

\item{match.to}{a numeric or character vector}
}
\value{
A numeric vector containing the row indices of every match, in the order that they appear in match.from
}
\description{
\code{matchIndices} is a function that returns row or element numbers where a set of values from one object occur in another.
}
\section{Warning}{

This function is not intended for direct use.
}

