% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnetwork.R
\name{nnetwork}
\alias{nnetwork}
\title{Find the number of distinct networks in a SpatialStreamNetwork}
\usage{
nnetwork(ssn)
}
\arguments{
\item{ssn}{an object of class SpatialStreamNetwork}
}
\value{
a single number
}
\description{
\code{nnetwork()} is a function to find how many separate networks there are in a single SpatialStreamNetwork object.
}
\examples{
\dontrun{NONE AS YET}

}
