% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nppoints.R
\name{nppoints}
\alias{nppoints}
\title{Calculate the number of prediction points on a network}
\usage{
\code{nppoints(ssn)}
}
\arguments{
\item{ssn}{An object of class SpatialStreamNetwork}
}
\value{
A list with two elements: a vector of prediction points by network, and a scalar being the total number of prediction points in the SpatialStreamNetwork object.
}
\description{
\code{nppoints()} is a function that returns the number of prediction points on a SpatialStreamNetwork object, both in total and by network.
}
\details{
To find the number of design points on a SpatialStreamNetwork object, use \code{\link{ndpoints}}.
}
