% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-method.R
\name{plot.ssndesign}
\alias{plot.ssndesign}
\title{Plot trace of the Greedy Exchange Algorithm}
\usage{
\method{plot}{ssndesign}(x, y,
  which.iteration = 1:length(x$trace.per.random.start), legend = TRUE,
  ...)
}
\arguments{
\item{x}{An object of class \code{ssndesign}}

\item{y}{Not used for \code{plot.ssndesign}}

\item{which.iteration}{A numeric indicating the trace that should be plotted. This can be a vector, in which case all traces corresponding to those indices will be plotted.The first five traces are plotted by default.}

\item{...}{Additional arguments to plot.}
}
\description{
A generic plot function adapted for `ssndesign` objects. It plots the traces from the optimisation algorithm.
}
\details{
The plot method produces line plots tracing the expected utility of the best design in the algorithm for each set of coordinate exchanges.
}
