% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateOnSSN2.R
\name{SimulateOnSSN2}
\alias{SimulateOnSSN2}
\title{Wrapper function for SimulateOnSSN}
\usage{
SimulateOnSSN2(...)
}
\arguments{
\item{...}{Arguments for \code{SSN::SimulateOnSSN}.}
}
\value{
A \code{SpatiaLStreamNetwork} object like the one returned in the \code{ssn.object} element of the list output from \code{SSN::SimulateOnSSN}.
}
\description{
This function was developed as a workaround for an error that occurs in \code{SSN::SimulateOnSSN} when a \code{SpatialStreamNetwork} object has temporal replicates on its sites.
}
