% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceFillingMaxiMin.R
\name{spaceFillingMaxiMin}
\alias{spaceFillingMaxiMin}
\title{A utility function for maximin space-filling designs}
\usage{
\code{spaceFillingMaximin(ssn, glmssn, design.points, prior.parameters, n.draws, extra.arguments)}
}
\arguments{
\item{ssn}{An object of class SpatialStreamNetwork. This argument may be ignored, but it is present so this function is consistent with all other utility functions.}

\item{glmssn}{An model object of class glmssn. Ignored in this function but this argument is present so this function is consistent with all other utility functions.}

\item{design.points}{A vector of pids corresponding to a set of observed sites in the obspoints slot of the SpatialStreamNetwork object.}

\item{prior.parameters}{A list of random functions that are parameterised in terms of n.draws.}

\item{n.draws}{A numeric scalar for the number of Monte Carlo draws to use when approximating the utility. Ignored in this function but this argument is present so this function is consistent with all other utility functions.}

\item{extra.arguments}{A list of extra parameters that control the behaviour of the utility function. The distance matrices required to compute covariance matrices are also stored in this list. Note that these are generated inside \code{\link{optimiseSSNDesign}}.}
}
\value{
A numeric scalar.
}
\description{
\code{spaceFillingMaximin} is a utility function used by \code{\link{constructSpaceFillingDesign}}. It is also fully compatible with \code{\link{optimiseSSNDesign}}. It is a utility function that maximises the minimum interpoint distance between a set of design points.
}
\details{
\code{spaceFillingMaximin} is deterministic and will ignore the arguments \code{prior.parameters} and \code{n.draws}. It is expected to be very fast, unless there are several hundred design points to select.
}
