% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSNdesign.R
\docType{package}
\name{SSNdesign}
\alias{SSNdesign}
\alias{SSNdesign-package}
\title{SSNdesign: An R package for optimal and adaptive sampling designs on stream networks}
\description{
The SSNdesign package builds on the SSN package to provide a framework for solving design problems on SpatialStreamNetwork objects.
}
\section{SSNdesign functions}{

the SSNdesign package contains the following functions:
- constructTotalMatrix
- CPOptimality
- doAdaptiveDesign
- doAdaptiveDesignParallel
- DOptimality
- EDOptimality
- EKOptimality
- findOptimalDesign
- findOptimalDesignParallel
- generateSites
- getBIDtables
- getImportantMatrices.obs
- getImportantMatrices.pxo
- getStreamDistMatsInOrder.predsxobs
- getStreamDistMatsInOrder
- glmssn_minimal
- hardCoreDesign
- importStreams
- KOptimality
- ndpoints
- nnetwork
- nppoints
- searchAllDesigns
- sequentialCPOptimality
- sequentialDOptimality
- sequentialEDOptimality
- SimulateOnSSN_minimal
- stableInverse
- systematicDesign
}

